+++ package.json
@@ -12,11 +12,15 @@
   "devDependencies": {
     "eslint": "6.1.0",
     "eslint-config-htmlacademy": "0.5.2",
+    "moment-locales-webpack-plugin": "1.2.0",
     "webpack": "4.42.1",
     "webpack-cli": "3.3.11",
     "webpack-dev-server": "3.10.3"
   },
   "engines": {
     "node": "10"
+  },
+  "dependencies": {
+    "moment": "2.25.1"
   }
 }
