+++ src/utils/render.js
@@ -21,21 +21,9 @@ const render = (container, component, place = RenderPosition.BEFOREEND) => {
   }
 };
 
-const replace = (newComponent, oldComponent) => {
-  const parentElement = oldComponent.getElement().parentElement;
-  const newElement = newComponent.getElement();
-  const oldElement = oldComponent.getElement();
-
-  const isExistElements = !!(parentElement && newElement && oldElement);
-
-  if (isExistElements && parentElement.contains(oldElement)) {
-    parentElement.replaceChild(newElement, oldElement);
-  }
-};
-
 const remove = (component) => {
   component.getElement().remove();
   component.removeElement();
 };
 
-export {RenderPosition, createElement, render, replace, remove};
+export {RenderPosition, createElement, render, remove};
